requirements.txt
-e git+https://github.com/jdupl/dhtxx-python3.git#egg=DHTxx
pyyaml
